+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - GDAL 3.3.0
   - NetCDF, GRIB, HDF5, Carto, FlatGeobuf, WMS, WMTS and WCS drivers
   - Support generalized strides in `gdal.RasterBandPixels.{write|read}()`
+  - Add `gdal.RasterBandPixels.{set|get}Async()`
 
 ### Changed
  - Linux release builds have lazy binding enabled
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Drop support for Node.js 10.x and 15.x
  - Drop support for Python 2
  - Drop support for the drivers removed in GDAL 3.3: AeronavFAA, BNA, E00Grid, HTF, OpenAir, SEGUKOOA, SEGY, SUA and XPlane
+
 ## [3.2.2] 2021-05-25
 
 ### Changed